Add -Wall to CFLAGS.  Warning fixes.
svn path=/trunk/; revision=1682

+++ src/core/librcscripts/debug.c
@@ -33,7 +33,7 @@
 static char log_domain[] = "rcscripts";
 
 void
-debug_message (const char *file, const char *func, size_t line,
+debug_message (const char *file, const char *func, int line,
 	       const char *format, ...)
 {
   va_list arg;
